Implemented recursive definitions in the interpreter
I had to rework the interpreter to get this all working

I was quite exasperated with some scoping problems, so I turned to the
wonderful 'Write You a Haskell' series, which suggests using explicit
closures. I'm a little tired to think about how it actually solved the
problem I was having, but it did.

Then there was the evaluation of recursive definitions- which WYaH did't
have any input on. I needed to store unevaluated code in the environment
as a Value (because if I tried to evaluate it to reduce to a value then
the program would run forever). So I figured that I could a "Pointer"
value to retreive the code and run it at will. And it works!